move declusterization to TMB
